pyproject.toml
[project]
name = "qdarchive"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10.0"
dependencies = [
    "python-dotenv>=1.0.1",
    "requests>=2.32.5",
    "pynada>=0.0.35",
    "pandas>=2.2.2",         
    "beautifulsoup4>=4.12.0"  
]